00_NIA_Emergency/requirements.txt
absl-py==1.3.0
astunparse==1.6.3
cachetools==5.2.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
et-xmlfile==1.1.0
flatbuffers==22.12.6
fonttools==4.38.0
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
importlib-metadata==5.2.0
joblib==1.2.0
JPype1==1.4.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
konlpy==0.6.0
libclang==14.0.6
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.24.1
oauthlib==3.2.2
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
Pillow==9.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
silence-tensorflow==1.2.1
SimpleITK==2.2.1
six==1.16.0
support-developer==1.0.5
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.1
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
xgboost==1.7.2
zipp==3.11.0
